+++ src/RelaxedIT.Update/RelaxedIT.Update.psm1
@@ -2,7 +2,7 @@
 
 
 function Test-RelaxedIT.Update {
-    Write-RelaxedIT -logtext "Test-RelaxedIT.Update v0.0.55"
+    Write-RelaxedIT -logtext "Test-RelaxedIT.Update v0.0.59"
 }
 
 function RelaxedIT.Update.All {
@@ -72,7 +72,9 @@ function RelaxedIT.Resources.Install {
     )
 
     # Define the modules to check and install
-    $modules = @("Az.Resources", "Az.Storage", "AzTable")
+    $modules = @("Az.Resources", "Az.Storage", "AzTable", "PSWindowsUpdate")
+
+    
 
     foreach ($module in $modules) {
         # Check if the module is installed
@@ -109,15 +111,20 @@ function RelaxedIT.Resources.OneclickInstall {
 
 function RelaxedIT.Update.Task {
     param (
-        [string]$LastrunTime = "C:\ProgramData\RelaxedIT\Update.Task.json"
+        [string]$LastrunTime = "C:\ProgramData\RelaxedIT\Update.Task.json",
+        [int]$writemmode = 1
     )
 
     try {
         Start-RelaxedLog -action "Update.Task"
 
+        if ($writemmode -ge 1) {
+           remove-item -Path $LastrunTime -ErrorAction silentlycontinue
+        }
         # Check if task should run using Compare-LastRun
         if (-not (Compare-LastRun -LastrunTime $LastrunTime -maxHours (7 * 24))) {
-        Write-RelaxedIT -LogText  "Task was executed less than 7 days ago. Skipping."
+            Write-RelaxedIT -LogText  "Task was executed less than 7 days ago. Skipping."
+            RelaxedIT.AzLog.Run.Ping -action "Skip"
             return
         }
         RelaxedIT.AzLog.Run.Ping -action "Start"