+++ test.ps1
@@ -1,0 +1,5 @@
+import-module .\src\RelaxedIT.Update\RelaxedIT.Update.psm1
+import-module .\src\RelaxedIT.3rdParty\RelaxedIT.3rdParty.psm1
+import-module .\src\RelaxedIT.AzLog\RelaxedIT.AzLog.psm1
+
+Test-RelaxedIT.Update